feat(util): Allows building for darwin targets
This is a simple change that allows the registrar to be built for MacOS.
All of the `unix` package being used works for Mac as well. The reason for
enabling this is completely selfish I'll admit. We are adding
[CSI Support](https://github.com/deislabs/krustlet/pull/462) to the Krustlet
project, which supports Mac as well. As most CSI implementations don't support
registering themselves with the Kubelet directly, we wanted to unblock any
users to need to register their plugins with a Krustlet Kubelet. This change
shouldn't affect normal container-based usage of the registrar